requirements.txt
matplotlib==3.5.2
numpy==1.21.5
pandas==1.4.4
plotly==5.9.0
python_dateutil==2.8.2
requests==2.28.1
snscrape==0.5.0.20230113
streamlit==1.17.0
streamlit_lottie==0.0.3
textblob==0.17.1
~treamlit==1.16.0
